actualizando endpoint cotizaciones

+++ src/main/java/movil/pos/venta/repository/entity/DetalleCotizacion.java
@@ -32,7 +32,7 @@ public class DetalleCotizacion implements Serializable{
     private Long productoId;
 
     BigDecimal precio;
-    BigDecimal cantidad;
+    int cantidad;
     BigDecimal subtotal;
     BigDecimal impuesto;
     BigDecimal descuento;
